compile_kernel: extract identifier/formula helpers to core.variable.utils
Move _sanitize, _assign_identifiers, _wrap_formula, _check_formula_arity and the
_RESERVED set into a new numbox/core/variable/utils.py so they have a stable,
reusable home independent of compile_kernel; compile_kernel imports the three it
uses back. Drop the now-orphaned re / keyword / inspect / CompileResultWAP
imports. Add the utils automodule to the variable docs and repoint the tests.
