+++ numbox/core/variable/compile_kernel.py
@@ -21,9 +21,6 @@ kernels never collide. A formula with no canonical fingerprint forces its unit
 uncached (no anchor, no numba cache) -- never reused, never wrong.
 """
 import hashlib
-import inspect
-import keyword
-import re
 import warnings
 
 from types import CodeType, FunctionType, ModuleType
@@ -35,13 +32,15 @@ from numba import njit, typeof
 from numba.core.ccallback import CFunc
 from numba.core.dispatcher import Dispatcher
 from numba.core.errors import NumbaError
-from numba.core.types.function_type import CompileResultWAP
 from numba.np.ufunc.dufunc import DUFunc
 
 from numbox.core.configurations import jit_options as _default_jit_options
 from numbox.core.variable._kernel_partition import (
     PartitionReport, Segment, _JitStep, _Plan, _PyStep,
     build_runs, discover, linearize, segment_liveness,
+)
+from numbox.core.variable.utils import (
+    _assign_identifiers, _check_formula_arity, _wrap_formula,
 )
 from numbox.core.variable.variable import (
     QUAL_SEP, CompiledGraph, CompiledNode, Graph, Variable, make_qual_name,
@@ -50,48 +49,8 @@ from numbox.utils.preprocessing import (
     _anchor_root, _materialize_anchor, _orphan_anchor_sweep,
 )
 
-# Names injected into the kernel exec namespace; identifiers must avoid them.
-_RESERVED = frozenset({"njit", "_kernel_jit_options"})
-
 _ANCHOR_SUBDIR = "numbox-compile-kernel"
 _orphan_anchor_sweep(_ANCHOR_SUBDIR)
-
-
-def _sanitize(qual_name: str) -> str:
-    s = re.sub(r"[^0-9A-Za-z_]", "_", qual_name)
-    s = re.sub(r"_+", "_", s).strip("_").lower()
-    if not s or s[0].isdigit():
-        s = "v_" + s
-    return s
-
-
-def _assign_identifiers(variables: list[Variable]) -> dict[Variable, str]:
-    """Map each Variable to a unique, valid, readable Python identifier.
-
-    Readable (from the qual_name) with a minimal deterministic sha256 suffix
-    only where names would otherwise collide. Reserves both the node temp `t`
-    and its formula global `f_<t>` so those namespaces never clash, and avoids
-    the injected reserved names.
-    """
-    used = set(_RESERVED)
-    idents = {}
-    for var in variables:
-        base = _sanitize(var.qual_name())
-        digest = hashlib.sha256(var.qual_name().encode("utf-8")).hexdigest()
-        cand = base
-        i = 0
-        while cand in used or ("f_" + cand) in used or keyword.iskeyword(cand):
-            i += 1
-            if i > len(digest):
-                raise RuntimeError(
-                    f"Cannot assign a unique identifier for {var.qual_name()!r}; "
-                    f"all sha256 prefixes exhausted"
-                )
-            cand = f"{base}_{digest[:i]}"
-        used.add(cand)
-        used.add("f_" + cand)
-        idents[var] = cand
-    return idents
 
 
 def _effective_flags(jit_options: dict | None) -> dict:
@@ -101,16 +60,6 @@ def _effective_flags(jit_options: dict | None) -> dict:
     as `fastmath`/`error_model` otherwise diverge across the discovery boundary)."""
     opts = {**_default_jit_options, **(jit_options or {})}
     return {k: v for k, v in opts.items() if k != "cache"}
-
-
-def _wrap_formula(formula: Callable, flags: dict | None = None) -> Dispatcher | CompileResultWAP | DUFunc | CFunc:
-    """Return an njit-callable for `formula`; plain-Python callables are njit-wrapped
-    with the kernel's effective jit `flags` so their semantics match the fused kernel."""
-    if isinstance(formula, (Dispatcher, CompileResultWAP, DUFunc, CFunc)):
-        return formula
-    if not callable(formula):
-        raise TypeError(f"formula {formula!r} is not callable")
-    return njit(**(flags or {}))(formula)
 
 
 class _Unfingerprintable(Exception):
@@ -233,21 +182,6 @@ def _formula_fingerprint(formula) -> tuple[str, bool]:
         return _fingerprint_function(target, set()) + extra, not isinstance(formula, CFunc)
     except (_Unfingerprintable, RecursionError):
         return f"{_safe_repr(formula)} @{id(formula)}", False
-
-
-def _check_formula_arity(formula, n_inputs: int, qual_name: str) -> None:
-    target = getattr(formula, "py_func", None) or getattr(formula, "__wrapped__", None) or formula
-    try:
-        sig = inspect.signature(target)
-    except (TypeError, ValueError):
-        return
-    try:
-        sig.bind(*range(n_inputs))
-    except TypeError as e:
-        raise ValueError(
-            f"{qual_name!r}: formula signature {sig} cannot accept its "
-            f"{n_inputs} declared input(s) passed positionally ({e})"
-        ) from None
 
 
 def _assemble_source(params: list[tuple[str, str, str]], lines: list[str], out_ids: list[str]) -> str:
